requirements.txt
pillow
numpy
pandas
opencv-python-headless
pytesseract